+++ src/Jadob/Router/Context.php
@@ -2,6 +2,10 @@
 
 namespace Jadob\Router;
 
+use function explode;
+use function strlen;
+use function substr;
+
 /**
  * TODO rename to RouterContext
  * @author  pizzaminded <mikolajczajkowsky@gmail.com>
@@ -26,6 +30,18 @@ class Context
     protected $port;
 
     /**
+     * There are at least two examples when alias will be filled:
+     *
+     * 1. App is hidden via Alias directive in webserver configuration
+     * 2. Document root has been run a level below our public directory
+     *
+     * @see https://httpd.apache.org/docs/2.4/mod/mod_alias.html#alias
+     * @see http://nginx.org/en/docs/http/ngx_http_core_module.html#alias
+     * @var string|null
+     */
+    protected ?string $alias = null;
+
+    /**
      * @return Context
      */
     public static function fromGlobals()
@@ -44,10 +60,30 @@ class Context
             return $context;
         }
 
-        $explodedHost = \explode(':', $host);
+        $explodedHost = explode(':', $host);
 
         $context->setHost($explodedHost[0]);
         $context->setPort((int)$explodedHost[1]);
+
+        //check for alias
+        //@TODO: check for CONTEXT_DOCUMENT_ROOT for Apache
+        //@TODO: check how does aliases are resolved in nginx
+
+        //URI requested by client
+        $requestUri = $_SERVER['REQUEST_URI'];
+
+        //trim query string
+        if(($questionMarkPosition = strpos($requestUri, '?')) !== false) {
+            $requestUri = substr($requestUri, 0, $questionMarkPosition);
+        }
+
+        //URI resolved by webserver
+        $pathInfo = $_SERVER['PATH_INFO'];
+
+        //if these two does not match that means that request has been "aliased"
+        if ($requestUri !== $pathInfo) {
+            $context->setAlias(substr($requestUri, 0, -strlen($pathInfo)));
+        }
 
         return $context;
     }
@@ -61,7 +97,7 @@ class Context
     }
 
     /**
-     * @param  string $host
+     * @param string $host
      * @return Context
      */
     public function setHost(?string $host): Context
@@ -79,7 +115,7 @@ class Context
     }
 
     /**
-     * @param  bool $secure
+     * @param bool $secure
      * @return Context
      */
     public function setSecure(bool $secure): Context
@@ -97,7 +133,7 @@ class Context
     }
 
     /**
-     * @param  int $port
+     * @param int $port
      * @return Context
      */
     public function setPort(?int $port): Context
@@ -105,4 +141,24 @@ class Context
         $this->port = $port;
         return $this;
     }
+
+    /**
+     * @return string|null
+     */
+    public function getAlias(): ?string
+    {
+        return $this->alias;
+    }
+
+    /**
+     * @param string|null $alias
+     * @return Context
+     */
+    public function setAlias(?string $alias): Context
+    {
+        $this->alias = $alias;
+        return $this;
+    }
+
+
 }