+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.67]
 ### Added
+- [Router] Partial support for aliased paths
 ### Changed
 ### Deprecated
 ### Removed